+++ History.md
@@ -1,3 +1,9 @@
+### 0.4.0 / 2014-08-10
+
+* added ability to use specified tags as dependency directives (not only `<link>` tags)
+* added ability to use shorter values for `rel` attributes (`css` instead of `stylesheet`, `include` instead of `x-include`
+and `require` instead of `x-require`)
+
 ### 0.3.0 / 2014-06-29
 
 * added ability to make inclusion depending on condition